Setup dataManage.js
