+++ src/web/app/src/components/SectionListPanel.vue
@@ -132,7 +132,7 @@
         return this.$store.state.dbMetaData.entitiesStatus.isLoading
       },
       versions() {
-        let list = Array.from(new Set(this.$store.state.presentation.versionList.map(v => v.versionId)));
+        let list = Array.from(new Set(this.$store.state.dataManage.versionList.map(v => v.versionId)));
         this.setDefaultValueForVersionList(list[0]);
         return list;
       },
