+++ src/web/app/src/store/index.js
@@ -8,6 +8,7 @@ import conference from "./modules/conference";
 import section from "./modules/section";
 import dbMetaData from "./modules/dbMetaData";
 import dataMapping from "./modules/dataMapping";
+import dataManage from "./modules/dataManage";
 
 Vue.use(Vuex);
 
@@ -29,6 +30,7 @@ export default new Vuex.Store({
     accessControl,
     section,
     dbMetaData,
-    dataMapping
+    dataMapping,
+    dataManage
   }
 })
