--- src/web/app/src/store/modules/data.js
@@ -1,7 +1,0 @@
-export default {
-    state: {
-        dataList: [],
-    },
-    mutations: {},
-    actions: {}
-}
